Added support for double-valued User points and improved points formatting with PointsDecorator, grouping and resizing label width to fit in big numbers

+++ QuizPlease/Modules/MainMenu/MainMenuPresenter.swift
@@ -26,7 +26,7 @@ protocol MainMenuPresenterProtocol: AnyObject {
     func didAddNewGame(with info: String)
     func didPressMenuRemindButton()
     
-    func userPointsAmount() -> Int?
+    func userPointsAmount() -> Double?
     func indexPath(for menuItemKind: MainMenuItemKind) -> IndexPath?
 }
 
@@ -99,7 +99,7 @@ class MainMenuPresenter: MainMenuPresenterProtocol {
         //
     }
     
-    func userPointsAmount() -> Int? {
+    func userPointsAmount() -> Double? {
         userInfo?.pointsAmount
     }
